requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.4
colorama==0.4.4
coverage==6.3.2
cryptography==36.0.1
Django==4.0.2
django-browser-reload==1.5.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
outcome==1.1.0
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.2.0
wsproto==1.0.0
